MI-99 User profile picture is retrieve from facebook based on facebook ID when loading profile page
