+++ mRides-app/UserProfileActivity.cs
@@ -13,6 +13,7 @@ using mRides_app.Models;
 using mRides_app.Mappers;
 using Android.Graphics;
 using System.Net;
+using Xamarin.Auth;
 
 namespace mRides_app
 {
@@ -35,12 +36,17 @@ namespace mRides_app
             usernameText = FindViewById<TextView>(Resource.Id.userName);
             usernameText.Text = user.firstName.ToString() + " " + user.lastName.ToString();
             List<Models.Feedback> userFeedback = UserMapper.getInstance().GetReviews(id);
-
-            ImageView facebookPicture = FindViewById<ImageView>(Resource.Id.userPhoto);
-            //var facebookPictureBitMap = GetImageBitmapFromUrl(User.currentUser.facebookPictureUrl);
-            // TODO: use the real link once the server is updated to contain the user profile picture
-            var facebookPictureBitMap = GetImageBitmapFromUrl("https://scontent.xx.fbcdn.net/v/t1.0-1/p50x50/13532894_10209745550106339_7260287654358119057_n.jpg?oh=70145db69abb22ca27097cb4f3d96e16&oe=5931D940");
-            facebookPicture.SetImageBitmap(facebookPictureBitMap);
+            
+            //var facebookPictureBitMap = GetImageBitmapFromUrl(user.facebookPictureUrl);
+            //TODO: use the real link once the server is updated to contain the user profile picture
+            UserMapper userMapper = UserMapper.getInstance();
+            var facebookPictureBitMap = userMapper.GetUserFacebookProfilePicture(user.facebookID);
+            if(facebookPictureBitMap != null)
+            {
+                ImageView facebookPicture = FindViewById<ImageView>(Resource.Id.userPhoto);
+                facebookPicture.SetImageBitmap(facebookPictureBitMap);
+            }
+            
 
             //For testing
             //FeedbackForNow fb = new FeedbackForNow();
@@ -68,27 +74,6 @@ namespace mRides_app
             feedbackListView.Adapter = userProfileFeedbackAdapter;
 
         }
-
-        /// <summary>
-        /// This method takes in the URL that goes to an image, and converts
-        /// the image into a bitmap to be returned.
-        /// </summary>
-        /// <param name="url">URL representing the image</param>
-        /// <returns>Bitmap of the image</returns>
-        private Bitmap GetImageBitmapFromUrl(string url)
-        {
-            Bitmap imageBitmap = null;
-
-            using (var webClient = new WebClient())
-            {
-                var imageBytes = webClient.DownloadData(url);
-                if (imageBytes != null && imageBytes.Length > 0)
-                {
-                    imageBitmap = BitmapFactory.DecodeByteArray(imageBytes, 0, imageBytes.Length);
-                }
-            }
-
-            return imageBitmap;
-        }
+        
     }
 }
